experiments/requirements.txt
groq
pandas
openai==1.55.3
httpx==0.27.2
langchain_community
transformers
fastembed
bert-score
torch
torchvision
torchaudio
nltk
rouge_score
langchain_qdrant
sentence_transformers
elasticsearch
evaluate
ollama